OCR_Transalator/requirements.txt
pillow
opencv-python
pytesseract
googletrans==4.0.0rc1
